moved input validation from repo to use case

+++ apps/frontend/src/app/features/manage-users/use-cases/delete.user.use.case.ts
@@ -2,13 +2,16 @@ import { Injectable } from '@angular/core';
 import { UserEntity } from '@clean-architecture-monorepo/shared';
 import { TranslatableError } from '../../../core/abstracts/translatable.error';
 import { UseCase } from '../../../core/abstracts/use.case';
+import { InvalidUserError } from '../manage.users.feature.errors';
 import { ManageUsersRepository } from '../repositories/manage.users.repository';
 
 @Injectable()
 export class DeleteUserUseCase implements UseCase<UserEntity, void> {
   constructor(private readonly repository: ManageUsersRepository) {}
 
-  execute(param: UserEntity): Promise<void | TranslatableError> {
+  async execute(param: UserEntity): Promise<void | TranslatableError> {
+    if (!param || !param.id) return new InvalidUserError();
+
     return this.repository.deleteUser(param);
   }
 }
